board: ti: am335x: update simple-framebuffer node.
In the case of simple-framebuffer, update its existing simple-framebuffer
node in the linux DT and enable it.

If the simple-framebuffer node is not detected by Linux kernel DT,
and the video is still active, then update the device-tree to
reserve this framebuffer.

This preserves the splash screen framebuffer till the display server
takes over once the OS is booted.

Signed-off-by: Sukrut Bellary <sbellary@baylibre.com>